requirements.txt
scons==4.8.1
ruff==0.6.4
pycryptodome==3.9.8
